Sync backend models with verified schema & fix auth screen scrolling
Detailed changes:
- Synchronized SQLAlchemy models (ElectricityBill, MeterReading, DeviceSession) with production PostgreSQL schema types and constraints.
- Enhanced Notification model and helper with context fields (account_number, bill_id, plan_id, action_url).
- Implemented IoTReading model for raw device logging.
- Fixed scrolling and layout visibility in RegisterScreen and LoginScreen.
- Optimized auth screen headers and spacing for better accessibility on mobile and web views.

+++ backend/member1-energy-analysis/src/models/budget_plan.py
@@ -92,7 +92,7 @@ class MeterReading(Base):
     
     # Calculated at time of reading
     units_consumed_so_far = Column(Integer, nullable=False)
-    days_elapsed = Column(Float, nullable=False)
+    days_elapsed = Column(Integer, nullable=False)
     actual_cost_so_far = Column(Float, nullable=False)
     expected_cost_so_far = Column(Float, nullable=False)
